Remove Setup.
At some point I want to add storage for the tracing so that trace data is persisted in the local WordPress install, at least temporarily so it can be viewed. . .but I need to put more thought into how to store it. This is a good first pass, but should come later as it’s own feature.