store-logs-in-wp
This is an attempt to store data in WP.

This definitely won’t scale as it is now. . .it _might_ work to schedule a cron and store the logs after the GraphQL request has ended, but right now this would tank the response time of GraphQL requests.

It’s a good initial stab, but I think we’ll really need to investigate offloading the requests to an external source, like Logstash / ES.

I think this should likely be a SaaS product that folks can pay for to add tracing to their WPGraphQL server. . .much like Apollo Optics, but specific for WPGraphQL.